% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{italian}
\alias{italian}
\title{Italian restaurants that are "less expensive"}
\format{A data frame with 66 observations of 6 variables
\describe{
\item{restaurant}{restaurant name}
\item{price}{price pf meal for two w/ drinks, in US dollars}
\item{food}{average food rating 0-30}
\item{decor}{average decor rating 0-30}
\item{service}{average service rating 0-30}
\item{east}{is the restauraunt east of 5th avenue?}
}}
\source{
http://gattonweb.uky.edu/sheather/book/docs/datasets/nyc.csv
}
\usage{
italian
}
\description{
New York Italian restaurants with an average price per dish under $40; eatin' good in the neighborhood!
}
\keyword{datasets}
