% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printemoji.R
\name{print_emoji}
\alias{print_emoji}
\title{This function prints a data frame with more cat fun}
\usage{
print_emoji(x, ...)
}
\arguments{
\item{x}{a dataframe}

\item{...}{arguements passed to \code{\link{print}}}
}
\description{
This function prints a data frame with more cat fun
}
